requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiofiles==0.4.0           # via quart
anyio==1.2.3
asks==2.3.6
async-generator==1.10     # via anyio, asks, trio
attrs==19.3.0
blinker==1.4              # via quart
boltons==20.0.0           # via face, glom
cachetools==4.0.0
certifi==2019.11.28       # via sentry-sdk
cffi==1.14.0              # via cryptography
click==7.0                # via quart
cryptography==2.8         # via pyjwt
face==20.1.1              # via glom
gidgethub==3.3.0
glom==19.10.0
h11==0.9.0                # via asks, hypercorn, wsproto
h2==3.2.0                 # via hypercorn
hpack==3.0.0              # via h2
hypercorn[trio]==0.9.0
hyperframe==5.2.0         # via h2
idna==2.9                 # via trio
itsdangerous==1.1.0       # via quart
jinja2==2.11.1            # via quart
marko==0.7.1
markupsafe==1.1.1         # via jinja2
outcome==1.0.1            # via trio
pendulum==2.0.5
priority==1.3.0           # via hypercorn
psycopg2-binary==2.8.4
pycparser==2.19           # via cffi
pyjwt[crypto]==1.7.1
python-dateutil==2.8.1    # via pendulum
pytzdata==2019.3          # via pendulum
quart-trio==0.5.0
quart==0.11.2             # via quart-trio
sentry-sdk==0.14.1
six==1.14.0               # via cryptography, python-dateutil
sniffio==1.1.0            # via anyio, trio
sortedcontainers==2.1.0   # via quart, trio
sqlalchemy==1.3.13
toml==0.10.0              # via hypercorn, quart
trio==0.13.0
typing-extensions==3.7.4.1  # via hypercorn
uritemplate==3.0.1        # via gidgethub
urllib3==1.25.8           # via sentry-sdk
werkzeug==1.0.0           # via quart
wsproto==0.15.0           # via hypercorn
